+++ dbx/utils/json.py
@@ -12,4 +12,6 @@ class JsonUtils:
 
     @staticmethod
     def write(file_path: Path, content: JsonContent):
+        if not file_path.parent.exists():
+            file_path.parent.mkdir(parents=True)
         file_path.write_text(json.dumps(content, indent=4), encoding="utf-8")